Fix a couple of Pwait() management bugs regarding dead processes.
If a process dies while it is stopped in a breakpoint handler, the DTrace
process control loop goes into a tight CPU-chewing loop, because the waitfd
reports that there is a process state change outstanding (the report of the
death), but Pwait() immediately returns rather than waitid() for a breakpointed
process (and possibly deadlock waiting for a process to return which cannot
execute).

Instead of immediately returning, simply do as is done for processes with
outstanding pending SIGSTOPs sent by libproc and refuse to do a blocking wait.
Nonblocking waits are quite safe in this situation and cannot deadlock.

However, one situation in which we *do* wait is unsafe: if we already know the
process is dead.  Waiting for it in this situation will usually fail, but if the
PID was recycled it may succeed, incorrectly, and might well block for a very
long time or even possibly deadlock.  In this situation, an immediate return
without waitid() *is* justified.