Do not build dtrace when check-installing.
The check-install family of targets is supposed to test the installed DTrace,
not a newly built one.  It is thus always a mistake for running this target to
build dtrace -- but after the USDT rules were installed, it does so in order to
run dtrace -G.

We should clearly run dtrace -G using the installed dtrace in this case, but
doing so is a little involuted, because we want the DOF rules to depend on
$(objdir)/dtrace if and only if make check-installed is not being run -- but
prerequisites are determined before target execution, not afterwards.

So we do a two-step thing.  Firstly, we set LD_LIBRARY_PATH and
DTRACE_OPT_SYSLIBDIR inside a wrapper script, $(objdir)/run-dtrace, and run and
depend on that in the DOF rules rather than directly on $(DTRACE).  We still set
DTRACE, but conditionally, only if unset, and the wrapper script depends on it.
The check-installed targets now recursively invoke make, setting DTRACE to
/usr/sbin/dtrace: to avoid an infinite recursion they invoke new
check-installed-*-internal targets which do what the corresponding
check-installed targets used to.  This recursive invocation scans all the
prerequisites from scratch, using the new command-line-provided value of
$(DTRACE), and thus avoids triggering a rebuild of $(objdir)/dtrace.