Do not explicitly specify installation user and group at install time.
Setting ownerships in our install targets meant we had to use fakeroot to do an
unprivileged install from RPM.

This is pointless: a simple install(1) without ownerships specified will get
the ownerships right, and on rpmbuild RPM sets the ownerships anyway.  This
removes our dependency on the fragile hack that is fakeroot (that stopped
working for me in my rpmbuild test environment).