unload-oops fixes.
Don't discard all the rmmod error output: it might be useful.

Don't even check if the dt_test rmmod worked: it might or might not in the
success case.

Sleep briefly before we end the test, to give the module time to reduce
refcounts in case this test was run alone and the testsuite harness is about to
unload everything (which will emit errors if the unload fails).

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>