Portability adjustments.
The ISA-dependent dispatch machinery unsurprisingly turns out to be inadequate
for real use, and the cross-platform breakpoint portability abstractions to be
insufficient.

So change the dispatcher machinery so that it uses a pair of macros to implement
each ISA dispatcher, so we don't have to rewrite the whole thing for every new
ISA-dependent function, and move the actual ISA-dependent functions from
isadep_$platform.c and isadep_impl.h to $platform/isadep.[ch].  This clashes
with the source file name for the dispatcher machinery, so rename that to
isadep_dispatch.c.

As for new dispatchers... we move the PTRACE_SINGLESTEPping into a new
ISA-dependent function, Pbkpt_singlestep(), which on SPARC will call back to do
a software singlestep.  It is unnecessary to adjust the IP address of the victim
before resuming breakpoints on SPARC, so we re-express that in terms of an
ISA-dependent Pget_bkpt_ip() that does all platform-specific IP address
adjustments, and an ISA-dependent Preset_bkpt_ip() which sets the IP address in
that one specific case (and on SPARC does nothing).  This means we don't have to
spend time on the completely different approach used on SPARC to set registers,
using regsets rather than PTRACE_POKEUSER.

The only generic-code-affecting change necessary for SPARC support in libproc
after this is software singlestepping.