Mark relocations for function entry addresses in DOF correctly.
The code generating the DOF ELF object was marking relocations for function
entry addresses using the R_X86_64_64 type, which works fine for executables,
but not for shared libraries.  Setting the type to R_X86_64_GLOB_DAT works for
both because it ensures that the relocation is resolved in both cases as the
absolute offset from the load address.  The DTrace core handles further
relocation based on the load address of the executable and/or library when the
DOF is loaded through the helper interface.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>