Fix deadlocks when doing things before dt_proc_continue().
When a process is started or grabbed and before dt_proc_continue() is done, the
DTrace process control thread is stuck on a condition variable waiting for the
continue to come in.  This is less than ideal, because the main thread often
sends proxy requests via a pipe to the process control thread: if it sends any
of these before dt_proc_continue(), DTrace will deadlock, because the process
control thread isn't listening to that pipe yet.

Fix by arranging for all communication from the main thread to the control
thread to go down the proxy pipe, including a new request indicating that a
dt_proc_continue() was just done.  Of course, this means we have to start
polling for requests inside dt_proc_stop(): so split out the entire polling loop
into a new dt_proc_loop() function, and call it from dt_proc_control() and from
dt_proc_stop() both.  Of course, if we're in the middle of a dt_proc_stop(), the
process is halted: we don't want to try to do a Pwait() on it, let alone risk
going into a breakpoint handler while we're already inside one: so
dt_proc_loop() takes a parameter causing it to *only* poll for incoming requests
from the main thread and not check the process for state changes at all.  (This
is safe, because while the process is stopped nothing much can happen to it
anyway, only signal receipt, and no harm is done if we put off transmitting
those signals until dt_proc_continue() time, since the process isn't running in
the interim anyway).  if it's in this 'main thread only' mode, it responds to an
incoming dt_proc_continue() by immediately returning, allowing dt_proc_stop() to
do just what it normally does.  If somehow this request comes in when
dt_proc_loop() has been called from the main thread, we just consume the request
and continue.  (This should be impossible, but it's easy to make it harmless if
it does happen, so let's do that.)

As a side benefit, the dpr_cv is now used to communicate in one direction only,
from the control thread to DTrace proper, so all the complexity and bugs around
dpr_cond_waiting and the need to consider that the dpr_lock is held even if it
looks like it isn't goes away.

Orabug: 17442388
Signed-off-by: Nick Alcock <nick.alcock@oracle.com>