Fix USDT header file cleanup.
The cleanup of USDT headers from inside the trigger Build file's
trigger-template wasn't working, because at the point at which this is expanded,
none of the per-trigger variable overrides have been defined yet.  One cannot
move that definition down to the end of the file, because a number of those
overrides depend on the trigger-template's variable definitions already having
been done, so they can override them.

So, instead, do the cleanup from the cmd-template, calling a new
hdr-clean-template for each relevant *_PROV definition, and having that add an
appropriate clean rule.