Software singlestepping support.
This adds a class of temporary breakpoints, signified by a non-NULL
bkpt_t.bkpt_singlestep, which connects back to their primary breakpoint: they
are dropped at the next instruction on platforms that do not support hardware
singlestepping when the time comes to singlestep. When hit, the primary
breakpoint's after-singlestep handlers are invoked instead, and the temporary
breakpoint is deleted. The temporary breakpoint's own handlers are ignored. If a
real breakpoint is dropped there, its handlers are executed in preference to the
after-singlestep handlers of the primary breakpoint. (DTrace will never do this,
since it only drops breakpoints at the start of functions.)

We don't support the case in which the next instruction equals the current one,
i.e. in which we would have to drop the temporary breakpoint on the very
instruction we're trying to singlestep over. This too is unlikely to be a
problem for DTrace's use case, but just in case it is we emit a message on
stderr before deleting the breakpoint.