Support platforms with no support for 32-bit binaries.
DTrace userspace has support for spying on the state of the dynamic loader in
32-bit binaries.  However, this support depends on the ability to compile and
link programs in 32-bit mode: if we can't do that, compilation will fail, which
is not ideal on platforms that don't support 32-bit binaries at all.

So verify that we can compile and link a 32-bit binary in the makefile, and if
we cannot, excise all 32-bit support, replacing relevant runtime loader offsets
with zeroes and arranging to emit an error and abort if asked to open a 32-bit
program.  We also avoid compiling 32-bit triggers in this case, and mark all
tests that depend on 32-bit support (such as those that depend on such triggers)
as expected to fail.

One potential new failure mode is that if you should attempt to compile DTrace
on a platform that has a 32-bit C library but no 32-bit development headers
installed, you will silently get a DTrace with no 32-bit support, which might
fail if asked to trace a 32-bit binary compiled on another platform.  The
specfiles should not do that, since they have 32-bit dependencies correctly
specified, but it is something to be aware of if compiling by hand.