Boost the timeout by a factor of ten when valgrinding.
Valgrind is sloooow, slow enough that fairly trivial tests take longer than ten
seconds: the docs suggest a slowdown of a factor of about ten (and a memory
usage consumption increase on the order of about two): this figure is quite old
and may even be too low in these days of deep memory hierarchies.  So boost all
timeout in this case.  The boost applies whether the timeout is the default of
10s, comes from --timeout, or from @@timeout.  (Hopefully this is not too
confusing.)