Assign the compressed CTF data to the right place.
When I introduced transparent CTF uncompression, I added a field in the
dt_module to hold the uncompressed content of the CTF... and then assigned the
uncompressed CTF to the *other* field, holding the CTF's name, leaking it.

This only went unnoticed because no use is made of that field after this point.