Fix dt_symbol_by_addr() symbol-type scanning loop.
The loop in dt_symbol_by_addr() is meant to scan backwards along the
dtst_addrs[] array of dt_symbols until it finds the first symbol with a given
address (since there may be multiple with the same address, and bsearch() is not
guaranteed to return any particular one).  Unfortunately, the searching loop was
completely messe dup, scanning backwards from a specific *element* of the array,
rather than in the array itself.  Since the elements are individually malloc()ed
this immediately fell into the void: it terminated at once when it did so (since
random garbage treated as a dt_symbol structure will probably have a different
dts_addr from what is wanted), but may return a non-optimal symbol in the
process.