Free the scopes array after constructing symbol search list.
The comments at the top of rd_get_scope() and rd_get_loadobj_link_map() both say
that the scopes array is dynamically allocated and persists across calls.  What
does its only caller do? That's right, it leaks it.

(Also fix said comments to get the name of the element to free right.)