Always remove breakpoints when Prelease() is called.
The breakpoint mechanism has for some time had a protective caul in place
preventing breakpoints from being removed when a removal is requested from
inside the breakpoint's handler (which is a convenient pattern to implement
one-shot breakpoints).  Instead, they are marked as pending removal, and removed
when the handler returns.

This normally works, but unfortunately fails in the presence of nonlocal exits.
For example, if DTrace is interrupted by a SIGINT right after requesting a
breakpoint removal inside one of its one-shot breakpoint handlers, and a
cancellation point happens between there and the return from the handler, thread
cancellation will be triggered and dt_proc_control_cleanup() will trigger a
Prelease() without ever exiting from the breakpoint handler.  Because in_handler
is still set on that breakpoint, it will not be removed: nor will the process be
resumed.  It's stuck, and even if DTrace dies it will still be stuck: if someone
SIGCONTs it, it'll still be in terminal trouble because there is still at least
one breakpoint dropped in it.

The obvious solution is to unconditionally flip off the in_handler flag when
flushing breakpoints.  This is always safe, because breakpoints are only flushed
in two situations: when the child process has exec()ed, or when Prelease()ing.
In the former case, the process must be running, thus we cannot be inside a
breakpoint handler.  In the latter case -- well, anyone who calls Prelease()
inside a breakpoint handler for the very process they're releasing deserves to
lose.  Don't do that.