Avoid leaking a lot of file_info structures on every purge.
purge_file_info() walks through the file_info list attached to the
ps_prochandle, deleting each in turn.  Unfortunately as a list counter it was
using P->num_files, which is decremented by the process of deletion, leading
to a number of file_infos not getting deleted at dtrace_close() time.

Fix trivial.