Use O_CLOEXEC in drti's open of /proc/$pid/maps.
We do this via the nonstandard glibc fopen() mode 'e', which is present in all
glibc's we care about (introduced in glibc 2.7).

This makes our /proc/$pid/maps fopen() safe in the presence of additional
concurrently-executing threads calling exec(), which is unlikely given our
constructor's priority but may still happen if this object is linked into a
shared library that is dlopen()ed later in a process's lifetime.

Jira: DTRACE-109