Split kernel module path lookup code into libdtrace/dt_kernel_module.c.
Finding a kernel module by name is quite difficult these days: you have to do a
fts() for it yourself.  The dt_kernpath_*() machinery to do this is private to
dt_module.c, but the new CTF module dumper could use it.  So split it out into a
new translation unit and mark it no longer static.

However, just because this is non-static doesn't mean we commit to its API: it
is very much a DTrace implementation detail, even though it must be exported
from the library if the CTF module dumper is to use it.  Add aliases to the
dt_kernpath_*() functions named dtrace__internal_*(), and make those public
under the symbol version LIBDTRACE_PRIVATE.  The combination of the name and
version should make it fairly obvious to anyone trying to use these symbols that
they're doing something wrong, and should make it almost impossible to collide
with their names by accident.