Multithreaded ustacks no longer crash.  Fix regex in tst.spin.sh.
The tst.spin.sh regex erroneously assumed that all ustacks were
inaccurate (ending in +something): this is not so.

Finally, arrange to print the DTrace output from tst.mtspin.sh to the test log:
without that, it's almost impossible to see why the test failed.  Also, we
don't want to return the exit status of a test ps(1) call: we want to return
the exit status of DTrace.

Orabug: 18412802