Fix libdtrace dependencies.
Make libdtrace self-contained, with no missing symbols pulled in by other
libraries. (This was always the intent, but mistakes had drifted in over time.)

This makes it possible to link tests with libdtrace's guts by depending on the
same set of libraries as libdtrace depends on, via $(libdtrace_LIBS).  So there
is no need to explicitly depend on any others any more.

Orabug: 18551552