dtrace: various fixes to tests
Enabling tests that previous would fail due to the absence of profile-n
probles.  Also fixing various tests that depend on the output of the
kernel stacktraces reported by stack().  A few small tweaks on top of
that for mistakes in tests that were imported from the original DTrace
testsuite implementation.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>