Dynamically search for l_searchlist in glibc's link map.
In order to reliably resolve symbols in the same way the dynamic linker
does (taking symbol interposition, -Bsymbolic, DT_GROUP and so forth into
account), DTrace has to acquire the search scope used to look up symbols
(we could do it in the current library, but currently we only use the main
program's search scope: see commit de8a4b3d1 for the rationale).  Unfortunately,
that scope is quite hard to obtain.  It is not in the public part of the link
map but a dozen fields deep, in the private "may change without notice" part --
and one of the fields that precedes it, l_info, is a giant array of pointers,
whose size depends on the number of ELF dynamic tags glibc currently knows
about.  So this offset changes frequently between glibc releases and sometimes
even within them. In this it is quite unlike all the other private
glibc-internal offsets rtld_db uses, mostly from the rtld_global structure,
which though they are private contain an admonition never to change the order of
the structure elements.

Before now, we were using the same mechanism for determining the location of
l_searchlist as for everything else: a compilation-time conversion of sizeof()
and offsetof() of a fake header containing definitions similar to glibc's link
map into a bunch of #defines (and a convenience scheme involving auto-filled
arrays which is not relevant here).  This is specific to one version of glibc --
the version used to manually derive the fake header -- so will not do.

So, instead, we can exploit the fact that we know what the symbol search scope
must contain -- a list of pointers to link maps, whose addresses we already
know.  So we enter a dynamic linker consistency region (to stop the link maps
changing out from under us), turn the addresses of the link maps into a sorted
array so we can bsearch() for them to determine if a possible pointer value
corresponds to a link map's address, then step across the link map hunting for
something looking like the guts of l_searchlist.  l_searchlist is an
r_scope_elem, the only such embedded directly in the link map:

struct r_scope_elem
{
  /* Array of maps for the scope.  */
  struct internal_link_map **r_list;
  /* Number of entries in the scope.  */
  unsigned int r_nlist;
};

So we look for an integer (greater than 1, since all primary link maps for
non-static executables contain at least two entries, for ld.so and the
executable) preceded by a pointer to an array of things that match our
already-known link map addresses.  If we find that, we've found the
l_searchlist.  If we fall off into unreadable addresses we can't find the link
map.  (Plus, as a sanity check, to avoid spending huge amounts of time
searching, if we get 64K from the start of the link map structure, we assume we
must have fallen off the end, and give up.  Since l_searchlist's current offset
is on the order of 700 bytes this seems safe enough.)

As for where to start searching -- we keep the glibc 2.12 version of the link
map around, and start looking for the l_searchlist at that offset.  We can
safely assume that DTrace will never be used with earlier versions of glibc, and
that the set of dynamic tags glibc supports will not fall: so it seems likely
that this offset will only increase with time.  As such, this is a useful
optimization, particularly given that glibc upgrades mean that many different
offsets may be in use in different processes on the system at the same time, so
we must carry out this scan afresh for every new process.