Look in the right place for architecture-dependent headers when substing .d's.
procfs.d was corrupted on some systems -- substitution of @ tokens in the .d.in
was not happening.  The reason for this is that the preprocessor was not finding
various crucial headers, all arch-dependent, and was failing before it emitted
the #defines we were looking for.

The solution is to look for these headers in the place they are found in the
kernel tree these days (under arch/x86/include).  We add a new variable in the
top-level GNUmakefile to track the name of this arch rather than hardwiring
'x86' into random places in the build system.

Orabug: 18652055