Do not allow symbol lookup, etc, on dead processes.
If a process dies while under monitoring, dt_proc_control_cleanup() will free
the rtld_db handle via a call to Prelease(), which calls Psym_release(), which
deletes the rd handle and sets it to NULL.

This all happens under the dt_proc_lock, but when the control thread dies the
lock is released. Operations blocked on said lock can now proceed.  They don't
know the process is dead and can proceed into Plookup_by_addr() and other
functions,

So arrange for all non-freeing-related public entry points in Psymtab.c to do
nothing when called on dead processes. (This is somewhat excessive -- some, e.g.
Pupdate_syms(), are only called from the control thread -- but doing it across
the board seems safer.)

(Ironically there used to be code in libproc to handle dead processes -- but it
was specifically to allow debugging of core dumps, which DTrace does not do, so
it was all torn out.)

(Also fix an indentation fubar in Pupdate_lmids().)

Orabug: 18550863