Close race window when process dies at the same instant as a proxy request.
The dtrace process control loop, dt_proc_loop() cycles through an endless round:

- release the dpr_lock
- wait for request via a waitfd
- take out the dpr_lock
- handle proxy request if one arrives
- do a nonblocking Pwait() if a state change has happened

The ordering is crucial: the main thread often does things (under the dpr_lock)
that cause a process state change, then requests a proxy Pwait().  So we must
handle proxy requests first.  However, if the process terminates on its own
(not under the lock) and our handling of that is immediately followed by a proxy
request, we never listen for that proxy request, thus can end up with the main
thread blocked on the dpr_cv and the process control thread dead.

In this situation, the main thread ends up waiting on the dpr_msg_cv condition
variable, whose state will never change because nothing is left alive to handle
it.

The obvious solution is to unconditionally signal the dpr_msg_cv when the
cleanup handler is run, after closing all the proxy pipes: if a proxy request is
outstanding, this will release it, and it is too late for any new ones to come
in (a proxy pipe write must always succeed before the cv is waited upon).

This problem is particularly common at process termination time, because process
termination involves a sudden surge of proxy calls as part of disconnection from
all monitored processes: if any die in this period, we hang.

Orabug: 18689795