Avoid array underrun when no textual mapping for the executable can be found.
If Pupdate_maps() fails to determine the textual mapping for an executable, it
sets P->map_exec (the array offset of said mapping) to -1, as a flag.  (Since it
is an array offset, 0 is an unsuitable flag value.)

Unfortunately in one place (map_iter()) we fail to check for that flag before
using it, leading to sucking out the contents of memory before the mappings
array as if it were the mappings array.  Even less fortunately we then
dereference a pointer out of it, which is of course garbage and causes a crash.
(Even less fortunately, some of the storage in this area is derived from the
contents of the executable, thus potentially attacker-controlled in small ways.)

Fix trivial: skip the mapping if it's the executable mapping and none was found.

Orabug: 18550863