KERNELDIR detection and header-substitution improvements.
The kernel header directory is needed to substitute preprocessing tokens in the
.d files that end up in /usr/lib64/dtrace.  It is not needed when these files
are not being built, whether because they already exist or because a
'make check-installed' is being carried out.

So check for the existence of the KERNELDIR only when the substitution is
happening, not unconditionally in the top-level makefile.

Further, the existing machinery to detect unsubstituted tokens is not
sufficient.  It works fine iff GCC continues with preprocessing even when errors
are found, but later versions of GCC sometimes abort processing completely when
#includes fail, leading to the following token-replace lines never being
emitted.  This yields an M4 script empty of the __sed_replace macros which
generate sed s///-expressions, which is then translated into an empty sed
script.  Since the sed script is empty, it doesn't contain any of the magic
@@@@SUBST-FAIL@@@@ tokens which indicate that an #define token was not found, so
no error is raised, even though substitution is sure to fail.

So, before M4 is run over the script produced by the cpp run, search it for
calls to __sed_replace.  If none are found, emit the stderr of the preprocessing
run and fail, so the user stands a chance of seeing what #include was failing
and what needs adjusting to make .d generation work again.