runtest: Do not log output while the timeout trap is in force.
This trap is on SIGCHLD, so fires for every subprocess termination.  It also
references a number of variables which are not set for the entire time the
trap is in force: in particular, the is some log output immediately after
the trap.  The preceding commit migrates log output to use a pipeline
ending in printf, which invokes and terminates subprocesses, firing the trap
before the variables are set and breaking the testsuite.

The fix is to migrate the trap-set to after the log invocation, and deactivate
it after we know its job is done.

Orabug: 19054052