dtrace: ensure that the DOF ELF object does not require execstack
The DOF ELF object was created without any indication on whether the
stack should be executable (no .note.GNU-stack section), and therefore
when it is linked with drti.o to produce the final ELF object that will
be linked in with an executable or shared library, that object is marked
as requiring an executable stack.  This then propagates into the final
executable.  Therefore, we now add an explicit .note.GNU-stack section
without the 'EXEC' flag set.

Orabug: 19217436

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>