Ensure that the test results are not polluted by mmap().
The expected results from this test are based on explicit operations
in the trigger program, but because the main probe that is used (for
map operations) may also be triggered by e.g. libc operations, i is
important to first record what fd we're interested in, and to then
use a predicate on the main probe to only report the operations we
are interested in.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>