Fix PS_STOP / PS_TRACESTOP confusion.
Now that Ptrace() loops waiting for PTRACE_INTERRUPT receipt, it is crucial that
we distinguish between the formerly almost identical process states PS_STOP and
PS_TRACESTOP.

PS_STOP means 'the process is stopped, but ptrace() is still listening for
signals on it, or will eventually switch into that state if Pwait() is called: a
PTRACE_INTERRUPT is meaningful'. PS_TRACESTOP means 'the process is stopped by
the tracer or stopped at an interrupt: PTRACE_INTERRUPT will deadlock.'

So arrange for breakpoint handlers that return PS_STOP to be silently
transformed into PS_TRACESTOP, and fix a few other places where loops were
mistakenly using the wrong state.