Stop background process monitoring around libproc calls in the main thread.
In normal operation, each process's monitor thread is listening for process
state changes via a waitfd and for incoming requests from code executing in the
dtrace master thread over the proxy pipe.  Unfortunately, this is vulnerable
to a class of nasty races.

To make it easier to reason about, libproc does not know about the existence of
multiple threads: all multithreading is handled via hooks which take out locks,
dispatch ptrace() calls and the like into the process control thread over the
proxy pipe, and the like.  A common pattern in libproc is to do something that
affects the process (like a PTRACE_INTERRUPT) and then do a blocking Pwait() for
the process to respond (by e.g. stopping): that Pwait() goes over the proxy pipe
and is ultimately serviced by the process control thread.

This pattern doesn't work very well if the process control thread is *already*
Pwait()ing: it is easy for the process control thread to consume the process
state change, following which the main thread and process control thread stall
waiting for a process state change that will never happen. (Technically this is
not exactly a deadlock, as you can rescue it by killing the traced victim, but
it's very close to one.)

So fix this by exploiting the wrappers recently introduced to handle execve()
restarting.  We introduce a new proxy request, proxy_monitor(), which turns
off (or on) the process control thread's background monitoring of process state
changes: if it is off, the process will only be Pwait()ed for if a proxy request
comes in asking for it: because it is itself a proxy request we don't need to
worry about interrupting the process control thread: it will automatically be
interrupted by the incoming proxy request and can note that it's been asked to
stop polling for process state changes.  We then call this new proxy_monitor()
function in the execve()-restarting wrappers.

The layering of the system now looks like this:

DTrace
  libproc wrappers (taking dtrace_prochandle_t *) handling reinvocation on
  exec(), and, now, disabling control-thread process-monitoring
    libproc
      proxied requests (e.g. ptrace() and Pwait())
        [thread transition into the control thread]
          process control loop
            ptrace(), Pwait(), breakpoint handler invocation, etc