Handle exec()ing processes correctly.
Currently, dtrace has a serious problem whenever a process exec()s.  It detects
the exec, resets some internal libproc state (including rtld_db state via the
exec-handler mechanism), then... returns to whatever it was doing, which might
well be halfway through dropping a breakpoint or something like that, relying
on dynamic state which just got free()d.  The result is, at best, a crash.

So replace this mess with an exception-handling mechanism whereby Pwait()
longjmp()s to a locus previously set up by one of its callers: that caller can
stash the previous value of the pad, recover, and longjmp() again up the
setjmp/longjmp chain.

By default, the chain root is an ordinary static variable in libproc.h, but this
is not sufficient for dtrace's use: libproc can be called into from both the
main thread and the process-control thread, and an exec() should affect both
without causing either to malfunction -- or, worse yet, to try to longjmp() into
the other thread, causing disaster.

So we allow libproc users to register a function which returns the address of a
jmp_buf *: the jmp_buf * can then be overwritten by callers to set up the pad.
We do this rather than the simpler approach of just registering a jmp_buf **
because dt_proc.c sets up a per-thread variable as the jmp_buf pad, and taking
the address of a per-thread variable loses its per-threadedness (the resulting
address points to only one thread's instance): calling a function that returns
the appropriate address every time it is needed ensures that you always get
the jmp_buf in the correct thread.

libproc also gains the ability to detach from a process without dropping the
ptrace() (via the new PS_RELEASE_NO_DETACH flag to Prelease()), and to attach to
an already-ptraced process (via the new Pgrab() already_ptraced flag).

dt_proc gains a dt_proc_reattach() function that uses these new facilities to
destroy and recreate the ps_prochandle inside a dpr_proc without detaching
from the process or allowing it to run.  We then soup up the newly-added
dtrace_prochandle proxying functions so that they intercept exec()s via the
new sjlj pad mechanism, and call dt_proc_reattach() and reinvoke the libproc
operation when this happens.  The net effect of all this is that a
dt_Pxlookup_by_name(), say, which is interrupted by its victim exec()ing, will
reconnect to the newly-exec()ed process, rescan its symbol table and dynamic
linker state, and redo the lookup, without its caller ever being aware of it.

The process control thread changes too.  It has to set up a reattaching longjmp
pad in two places: firstly, one that traps exec()s during most of its execution
lifespan, and secondly an inner one which traps exec()s which are happening
while handling a proxy request from the master thread.  If this is seen, it
responds by setting a new dpr_proxy_exec_retry flag, which the proxies detect
and transform into a longjmp() along *their* thread's sjlj pads.  Thus we proxy
a detected exec()'s exception throw just as we proxy a normal function return.

Finally, the process of reattachment itself is one that can only be performed by
the thread that's attached in the first place, the process control thread: so
dt_proc_reattach() is itself proxied to the process control thread.

So, if an exec() happens inside, say, a breakpoint deletion which happens
as a result of a breakpoint trap which was fired while Pwait()ing while
setting up an rtld_db debugger instance while doing a proxied Plookup_by_addr()
from inside dt_aggregate_usym() (a failure mode I have actually seen), we
will see the following:

-- master thread --
dt_aggregate_usym()
dt_Plookup_by_addr() (longjmp() pad 1)
Plookup_by_addr()
Pupdate_lmids()
rd_new()
Pwait() (calls wrapper)
proxy_pwait()
-- process control thread --
/* ... */
dt_proc_loop() (longjmp() pad 2)
proxy_pwait() (internal call, no longjmp pad)
Pwait_internal()
Pwait_handle_waitpid()
  -- exec detected!

When this exec is detected, we jump out to pad 3, which notes that an exec was
detected and does a cross-thread proxy notification to its caller.  This
then throws to the next longjmp() pad (rd_new() is careful to do its only
Pwait() before it does anything that needs cleanup, so it doesn't need one),
which is the dt_Plookup_by_addr() wrapper.  That calls dt_proxy_reattach(),
which makes another cross-thread proxy call to the process control thread,
destroys the ps_prochandle without detaching from it, grabs it again without
ptracing, and calls Puntrace() to get it out of PTRACE_O_TRACEEXEC, which it's
been stuck in ever since the exec() was detected.  Finally we drop back out
to dt_Plookup_by_addr(), which reseats the ps_prochandle inside its
dtrace_prochandle wrapper and tries the function call again.

We add a libproc internals testcase that drops breakpoints on a process that is
continuously exec()ing, and makes sure that this works and that its sjlj pad is
thrown to sometimes.  We also add a unittest/proc testcase that dtraces a
continuously-exec()ing process and makes sure that dtrace doesn't lose control
of it, and a test that does something similar with an active pid
provider.  (This is skipped for now because it passes despite the lack of a pid
provider, causing an XPASS.)

There are some delicate and unpleasant aspects to all this, as is always true of
exception handling in C.  The most obvious part, and the easiest part to fix, is
the ugly repetitious wrappers for libproc calls.  These will be replaced in
future with machine-generated code, so we can ignore them for now.  The
unpleasant parts we can't ignore are twofold: firstly, all future changes to the
dt_proc process-control thread code, and to anything in libproc at all, must be
done with the understanding that the victim process could exec(), and the
function could be longjmp()ed past: working around this and setting up a longjmp
pad means making all variables in the current function which are live across
that pad volatile, and piling in ugly casts back to non-volatile wherever they
are used.  These problems are, I fear, unavoidable.

Orabug: 19046684