Fix race causing crash when initial attachment or grabs fail.
When a process dies (or, in theory, when some other interesting event happens --
but we have no other interesting events at present), the process control thread
cleanup function enqueues a notification onto a linked list and wakes up the
main thread out of dtrace_sleep() so it can note the death of the process. The
notification contains the dt_proc_t associated with the ex-process, so the main
thread can find out what went wrong. (Lifetime issues are resolved by the dph
hash: dead entries on it are expired by the main thread, long after notification
has finished.)

Normally, this works fine, but if initial attachment or creation fails severely
enough that the process is never created, the main thread notes this in
dt_ps_proc_{grab,create}() and free()s the dpr, leaving a wild pointer enqueued
on the notification list, which causes crashes when, later, the main thread gets
around to sleeping and checking for notifications.

The solution is for the process control thread cleanup function to enqueue the
dpr only when the process already existed (and thus was safely disconnected):
otherwise, NULL is enqueued.

Orabug: 19679998