Fix a pair of connected problems causing crashes when exec() was detected.
Firstly, our attempt to declare relevant variables volatile around longjmp()
landing pads was stymied by C delarator syntax: the 'volatile' was in the wrong
place, leading to non-volatile pointers to volatile data rather than vice versa.
With that fixed, all the horrible casts can be removed again, too.

Secondly, we were failing to note that the when the unwinding happens, we don't
reset the landing pad at every stage, we just fish the next stage of the
unwinding out of our local automatic variables and jump on. So the first thing
we do in the successful-setjmp() case *must* be to reset the landing pad, since
it's still pointing to some setjmp() which has already been unthrown, way down
in the bowels of libproc.

Thirdly, a couple of testcases were failing to preserve all their automatic
variables across in the setjmp()-successful case.

Orabug: 19046684