Erase process-death notifications left dangling by failed process creation.
When a process being monitored by dtrace userspace dies or changes state, a
notification is enqueued so that the dtrace main thread can observe this fact
and call the dtrace_prochandler so that the libdtrace consumer can be told.
These notifications may not be collected if dtrace_sleep() is never called
again, so something else must clean them up.  Normally this is
dt_ps_proc_destroy(), which drops all notifications relating to the process it's
been asked to free up all sign of; but if a process dies before its controlling
thread is successfully created and rendezvoused (i.e. fork(), exec(), or initial
attachment fails), we never free it because dt_ps_proc_destroy() is never called
on it.

So free any remaining such notifications in dt_proc_hash_destroy(): if the hash
is vanishing, all the notifications enqueued on it must necessarily vanish too,
since they will never be referenced again.

Orabug: 19679998