Load modules from /etc/dtrace-modules when modprobing for dtrace.
Comments are supported (leading #, as usual).

Probing for these modules is nonfatal, and stderr is discarded, meaning that not
all modules need to be installed: those that are installed will work, and those
that are not will simply not contribute probes to the available set.

This means we can ship an /etc/dtrace-modules that contains all modules we
consider usable, from in-tree, out-of-tree, anywhere, and have them autoloaded
if they are available.

This probe only happens when the dtrace module itself is not loaded: i.e.,
usually, on first dtrace run after boot.

Orabug: 19821254