runtest: Fix support for quiet unloading.
We used to implement a @@quiet tag which told the testsuite that if a module
could not be unloaded, it was not an error (thus supporting kernels in which
a module required for testing was built in).

Unfortunately, we broke this recently when transitioning to a dtrace-loaded
scheme.  While the unload-only tag should indeed be implemented entirely as a
comment (to prevent dtrace from loading it), the unload warning tag should be
implemented as a *trailing* comment, since dtrace has to load it as usual: only
the testsuite's autounloading code should know that it is special.

This involves a change to load_dtrace_modules, too, since it now has to support
trailing comments on a line, not just comments covering the whole line.

While we're about it, rename the tag to 'unload-quietly' to make it clear that
this only affects unloading.

Fix trivial.

Orabug: 19821254