Use appropriate kernels to build per-kernel D libraries.
The previous change caused us to place D files in per-kernel subdirectories of
the DTrace directory -- but each of these was still being generated by
#including headers from the same place, a single kernel include tree set on a
per-build basis.

Fixing this is fairly simple.  Rather than pass in the path to a kernel build
tree, we split the path into three pieces, the part before and after the kernel
version (which is invariant) and the kernel version itself, and we allow callers
to pass in a space-separated string of kernel versions, e.g.

make KERNELS="3.8.13-44.1.3.el6uek.x86_64 3.18.0-1.el6uek.x86_64"

As before, we default these to $(uname -r) and a suffix such that the
/lib/modules/$version/build link is searched, and override them in the RPM
specfile.  Then the libdtrace Build file slices these apart using awk to derive
the short names used in the dlibs directories (much as DTrace itself does using
strchr() at runtime), and passes both the short and long forms into the
M4-processing target, which can then use them.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>