Fix exec()ing inside a dynamic linker consistency region.
If an exec() happens when we are inside a dynamic linker consistency region, we
were throwing all the way out of it and then invoking
Prelease(PS_RELEASE_NO_DETACH) and a Pgrab() with already_ptraced set, to throw
away our local state and reacquire the process without allowing it to continue
execution (so that we miss nothing while getting our house in order).

Unfortunately, one of the things Prelease() must discard (by way of
Psym_release()) is rtld_db state (by calling rd_delete() if an rtld_db is active
on the process) -- and this checks to see if we are in a dynamic linker
consistency region, and repeatedly calls rd_ldso_consistent_end() until it ends.
This influences the state of the child process, removing various breakpoints
(which are, of course, already gone, since we've exec()ed) and does Puntrace()s
and Pbkpt_continue()s (which themselves do Puntrace()s) to resume the process as
many times as necessary.  This is exactly what we were trying to avoid: the
process is set going long before we're ready for it, and because it's now running
but libproc thinks it is stopped, everything we do at initialization time (e.g.
dropping dynamic linker breakpoints) fails with ESRCH, so we lose track of the
process's dlopen()s and dlclose()s too.

In normal dtrace use this failure mode is probably quite rare -- the victim
would have to exec() right after a dlopen() or dlclose() -- but it's perfectly
possible.  The internals/libproc/tst.consistency.sh test triggers it often.

To fix this, we introduce a new function (rd_ldso_consistent_reset()) which
resets enough of the rtld_db state that destroying it won't try to do any
rd_ldso_consistent_end()s, and will thus leave the state of the process alone:
then we ensure that all callers of the dynamic linker consistency enforcement
functions also set up exec()-detection longjmp pads that call
rd_ldso_consistent_reset() if an exec() is detected.  (Since the
rd_ldso_consistent_*() functions are not on the stack while we are in a
consistency enforcement region , we cannot drop an exec()-detection longjmp pad
in the *_consistent_end() functions to do this job, but must drop it in their
callers instead.  This is not really a problem: there are two *_consistent_end()
functions, and two callers of these functions, and one of those has a pad
anyway, we're actually saving a little code by being forced to stick the job
with the callers!).