Disable dlopen() monitoring in multithreaded processes.
This is a terrible hacky workaround for lack of multithread-capable
libproc (since getting the bugs out of that is proving tricky).  Instead, we
monitor thread creation operations and disable rtld event monitoring whenever a
new thread creation event is seen.  We never re-enable it again: the ratchet is
one-way, for simplicity's sake (and because anything else would suddenly require
us to count threads and determine when there was only one left and import a huge
chunk of the libproc-mt branch's state tracking machinery).  Breakpoints can
still be dropped to maintain dynamic linker state consistency, but these are
only held for very brief periods and thus are correspondingly unlikely to cause
crashes (though, alas, it can still happen -- it's much better than it used to
be, though).

This means that symbol lookup inside dlopen()ed libraries will still work for
everything loaded while a program was single-threaded: all the other stuff we
can do with dynamic linker state monitoring, even ELF symbol scope-sensitive
name lookup et al, will work in multi-threaded programs (except if the symbol in
question was introduced by a dlopen() after the program went multithreaded).

This is much better than what we had before, and a tiny change compared to the
huge destabilizing multithread-capable libproc change.

Orabug: 21279300