Wait for state changes in (v)fork()ed / clone()d ptrace()d children properly.
When a forked or cloned child of a process under ptrace() is created under
tracefork / traceclone control, the parent stops with a PTRACE_EVENT to
note that this has happened.  Unfortunately, the assumption I made that
by the time this has happened the child is stopped is not true: a stop
is sent to it using the normal signal-handling mechanisms, so not only
can the signal-receipt be scheduled after the ptracer executes and
receives the ptrace event (so that it's not ready to be traced itself),
it can even receive other signals first! If the system is busy enough,
it can even execute code, fork again, or exit before it gets seized.

So there are three cases when a tracefork / traceclone is received

 - the child is already in traced state
 - the child is not in traced state, but will be shortly
 - the child is not in traced state, and other things will happen before
   it is ready

Unfortunately, because the tracefork and traceclone options are inherited by
child processes, this means we must be ready to handle not only signal
dispatch (PTRACE_CONTing the child until such time as the queued interrupt gets
through) but we must even be ready to handle the child doing further forks or
clones, and then we have to do the same thing to *those* children.  There are
further complications around exec(): if one of those trickles in while we're
waiting for a post-fork()/clone() interrupt, we still have to wait for the
interrupt to happen, but we *don't* have to do things like removing old
breakpoints any more, because we now know that the breakpoints must have been
wiped out by the exec().

There is, as yet, no testcase for the 'fork()-and-then-signal-or-exec' cases:
they are incredibly hard to trigger, taking hours of test runs. (The reported
bug was simply a case of the ptracer getting scheduled before the ptraced child,
and even *that* is rarely triggered and requires extra load on my test systems
before it will happen at all.)

As an additional closely-related change, we rejig things so that on a clone(),
we turn PTRACE_O_TRACECLONE off: we only use it to flip a 'this process has
>1 threads in it' latch: it's just a waste of CPU time to report all future
thread creations, since we don't care about any of them after that.

Orabug: 21284447