Drop the dtrace, libdtrace and testsuite version into the test logs.
We use the version-reporting infrastructure in the previous commit to
report the precise version of dtrace(1) and libdtrace in use, in addition
to sucking the information out and using it to report the testsuite
version as well. (This seems like it should be redundant, but once we
have a working installable testsuite, it won't be.)

This is rendered a little more complex by the desire to have this work in
autobuilt source tarballs too.  We do this by having the 'git archive' output
come from git direct if and only if this is a git checkout: otherwise, it is
copied from .git-archive-version, which is a file subject to
.gitattributes-driven export substitution.  (We add a dist target at the same
time, since we are now making the make-a-tarball workflow official enough to
have makefile machinery to handle it, and since we do this routinely to feed the
autobuilders.)