Add a testsuite RPM.
This is simple enough, just another subpackage using the just-added installable
testsuite machinery, but there are two complexities: firstly, unlike most of
dtrace, which contains no debugging information worth mentioning, this package
contains files which *must not* be stripped if the tests are to pass, and
secondly it contains shared libraries in an unusual location which perturbs the
automatic prerequisite detection code.

So turn both debuginfo generation and prereq detection off, the first for the
entirety of dtrace (since turning it off for only one subpackage is really
quite difficult), the second for only the testsuite subpackage.

(There is a *bit* of debugging info in dtrace, but it crept in from the startup
.o files.  It's useless, so we now strip that out too.)